Module Resto_cohttp_client.Client

Client calls to services.

module type CALL = sig ... end

The minimal interface for building a client. Cohttp_lwt.S.Client is an instance of this signature, modulo some additional optional parameters that resto does not use. See OfCohttp below to obtain an exact intance based on Cohttp_lwt.S.Client.

module OfCohttp (Client : Cohttp_lwt.S.Client) : CALL

module Make (Encoding : Resto.ENCODING) (Call : CALL) : sig ... end

Make(Encoding)(Client) is a module that allows you to make calls to various Resto (or EzResto) services.