Module type Client.CALL

The minimal interface for building a client. Cohttp_lwt.S.Client is an instance of this signature, modulo some additional optional parameters that resto does not use. See OfCohttp below to obtain an exact intance based on Cohttp_lwt.S.Client.

val call : ?⁠headers:Cohttp.Header.t -> ?⁠body:Cohttp_lwt.Body.t -> Cohttp.Code.meth -> Uri.t -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t